First completed version of multiping.  Uses upper four bits of the ICMP sequence number field in the ICMP header to keep track of which remote site the echo is going to or coming from.
This version was reconstructed backwards, after-the-fact, so it probably
has some bugs and quirks in it.  Things to check are the bit manipulation
for the sequence ID numbers, and whether or not prefinish() screws things
up.  The only other changes to be made are to make a separate usage()
function and a output_new_style() function for clarity.